rename swagger package to importer
