+++ sysl2/sysl/importer/import.go
@@ -1,4 +1,4 @@
-package swagger
+package importer
 
 import (
 	"bytes"
