+++ sysl2/sysl/importer/indentwriter.go
@@ -1,4 +1,4 @@
-package swagger
+package importer
 
 import (
 	"io"
