+++ sysl2/sysl/importer/indentwriter_test.go
@@ -1,4 +1,4 @@
-package swagger
+package importer
 
 import (
 	"bytes"
