+++ sysl2/sysl/importer/parameters.go
@@ -1,4 +1,4 @@
-package swagger
+package importer
 
 import (
 	"fmt"
