+++ sysl2/sysl/importer/tests-swagger/EXAMPLE_SWAGGER_SPEC_WITH_ENDPOINT_PATH.yaml
