+++ sysl2/sysl/importer/tests-swagger/EXAMPLE_SWAGGER_SPEC_WITH_ENDPOINT_PATH_WITH_ERROR_RESPONSE.sysl
