+++ sysl2/sysl/importer/tests-swagger/EXAMPLE_SWAGGER_SPEC_WITH_ENDPOINT_RETURNING_ARRAY_OF_DEFINED_OBJECT_TYPE.yaml
