+++ sysl2/sysl/importer/tests-swagger/SIMPLE_SWAGGER_EXAMPLE.sysl
