+++ sysl2/sysl/importer/tests-swagger/SWAGGER_HEADER_AND_BODY_PARAM_EXAMPLE.sysl
