+++ sysl2/sysl/importer/tests-swagger/SWAGGER_OBJECT_WITH_A_REQUIRED_PROPERTY.yaml
