+++ sysl2/sysl/importer/tests-swagger/SWAGGER_OBJECT_WITH_NO_PROPERTIES.sysl
