+++ sysl2/sysl/importer/tests-swagger/SWAGGER_QUERY_PARAM_EXAMPLE.sysl
