+++ sysl2/sysl/importer/tests-swagger/SWAGGER_REQUIRED_AND_OPTIONAL_FIELDS_EXAMPLE.sysl
