+++ sysl2/sysl/importer/tests-swagger/SWAGGER_TOP_LEVEL_ARRAY_EXAMPLE.yaml
