+++ sysl2/sysl/importer/tests-swagger/SWAGGER_WITH_ARRAY_IN_OBJ.yaml
