+++ sysl2/sysl/importer/tests-swagger/SWAGGER_WITH_HEADER_VAR_OVERRIDDEN_IN_METHOD.yaml
