+++ sysl2/sysl/importer/tests-swagger/SWAGGER_WITH_PATHS_VAR_REFERRING_GLOBAL_PARAMS_OBJECT.sysl
