+++ sysl2/sysl/importer/tests-swagger/SWAGGER_WITH_PATH_VAR_TYPE_OVERRIDDEN_IN_SECOND_METHOD.sysl
