+++ sysl2/sysl/importer/tests-swagger/SWAGGER_WITH_SYSL_KEYWORDS_EXAMPLE.sysl
