+++ sysl2/sysl/importer/tests-swagger/SWAGGER_WITH_TYPELESS_ITEMS.sysl
