+++ sysl2/sysl/importer/types.go
@@ -1,4 +1,4 @@
-package swagger
+package importer
 
 import (
 	"fmt"
